requirements.txt
opencv-python>=4.8.0
pytesseract>=0.3.10
numpy>=1.24.0
Pillow>=10.0.0
supabase>=2.0.0
python-dotenv>=1.0.0
setuptools>=65.0.0
